refactor(keyboard-manager): separate testing components from implementation
