+++ src/index.tsx
@@ -1,3 +1,3 @@
 export * from './keyboard';
-export * from './types';
 export * from './simulate-key-down';
+export * from './types';
