+++ src/keyboard.tsx
@@ -1,18 +1,17 @@
 import * as React from 'react';
 import * as uuid from 'uuid';
-import { TestKeyDown, RegisterKeyDownTestProps, KeyboardManagerProps } from '.';
-
+import { KeydownEv, KeyboardManagerProps } from '.';
 
 export class KeyboardManagerContext {
   public readonly id = uuid.v4();
-  private testKeyDowns: TestKeyDown[] = [];
+  private keydownArr: KeydownEv[] = [];
   private document?: Document;
 
-  public keyDown = (ev: KeyboardEvent) => {
-    this.testKeyDowns.forEach(t => t(ev));
+  public keyDown = (ev: KeyboardEvent): void => {
+    this.keydownArr.forEach(t => t(ev));
   };
 
-  public injectDocument(document?: Document) {
+  public injectDocument = (document?: Document) => {
     if (document && this.document !== document) {
       if (this.document) {
         this.document.body.removeEventListener('keydown', this.keyDown);
@@ -22,16 +21,16 @@ export class KeyboardManagerContext {
     }
   }
 
-  public registerKeyDownTest(t: TestKeyDown) {
-    this.testKeyDowns.push(t);
+  public registerKeyDownTest = (t: KeydownEv) => {
+    this.keydownArr.push(t);
   }
 
-  public unregisterKeyTest(t: TestKeyDown) {
-    const found = this.testKeyDowns.indexOf(t);
+  public unregisterKeyTest = (t: KeydownEv) => {
+    const found = this.keydownArr.indexOf(t);
     if (found < 0) {
       throw new Error('Try unregister does not exists');
     }
-    this.testKeyDowns.splice(found, 1);
+    this.keydownArr.splice(found, 1);
   }
 }
 
@@ -41,36 +40,6 @@ const KeyboardManagerCtx = React.createContext<KeyboardManagerContext>(
 );
 
 export const KeyboardManagerConsumer = KeyboardManagerCtx.Consumer;
-
-
-interface InternalRegisterKeyDownTestProps extends RegisterKeyDownTestProps {
-  readonly keyboardManagerContext: KeyboardManagerContext;
-}
-class InternalRegisterKeyDownTest extends React.Component<
-  InternalRegisterKeyDownTestProps
-> {
-  public componentDidMount() {
-    this.props.keyboardManagerContext.registerKeyDownTest(this.props.testFn);
-  }
-  public componentWillUnmount() {
-    this.props.keyboardManagerContext.unregisterKeyTest(this.props.testFn);
-  }
-  public render() {
-    return <>{this.props.children}</>;
-  }
-}
-
-export function RegisterKeyDownTest(props: RegisterKeyDownTestProps) {
-  return (
-    <KeyboardManagerConsumer>
-      {ctx => (
-        <InternalRegisterKeyDownTest {...props} keyboardManagerContext={ctx}>
-          {props.children}
-        </InternalRegisterKeyDownTest>
-      )}
-    </KeyboardManagerConsumer>
-  );
-}
 
 export function KeyboardManager(props: KeyboardManagerProps) {
   keyboardManagerContext.injectDocument(props.document || document);
